% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/airsea-package.r
\docType{package}
\name{airsea}
\alias{airsea}
\alias{airsea-package}
\title{airsea.}
\description{
package for doing stuff with the sea and the air
}
\references{
Hull et al (2015)

Johnson, M. T. (2010)
}

