% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liang.R
\name{B_L12}
\alias{B_L12}
\title{Liang et al 2012 Bubble param}
\usage{
B_L12(u10, sst, gasname, sal)
}
\arguments{
\item{u10}{}

\item{sst}{}

\item{gasname}{}
}
\value{
bubble sat term (dimensionless)
}
\description{
Liang et al 2012 Bubble param
}

