% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gas.R
\name{bubbleSat}
\alias{bubbleSat}
\title{Equilibrium bubble supersaturation}
\usage{
bubbleSat(compound = "O2", u)
}
\arguments{
\item{compound}{character string of compound of interest ('O2' ,'N2', 'Ar' or 'CO2')}

\item{u}{vector of wind speed in meters per second at 10 meters height}
}
\value{
vector of equilibrium bubble supersaturation effects (no units)
}
\description{
Implements the bubble supersaturation parametrisation of Woolf and Thorpe (1991)
Note: parametrisation is only applicable for Oxygen (O2), Nitrogen (N2), Argon (Ar) and Carbon Dioxide (CO2).
}
\details{
TODO
}
\examples{
bubbleSat('Ar', 7)  # bubble supersaturation effect for Argon at 7 m-1 s-1 winds
}
\references{
Woolf, D. K., & Thorpe, S. A. (1991). Bubbles and the air-sea exchange of gases in near-saturation conditions. Journal of Marine Research, 49(3), 435-466. doi:10.1357/002224091784995765
}
\keyword{bubble}
\keyword{exchange}
\keyword{gas}
\keyword{kw}

