% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_dhdt}
\alias{calc_dhdt}
\title{Rate of change in mixed layer}
\usage{
calc_dhdt(h0, h1, time, entrainmentOff)
}
\arguments{
\item{h0}{the mixed layer depth at time 0}

\item{h1}{the mixed layer depth at time 1}

\item{time}{a vector containing the maximum time interval}
}
\value{
the rate of change in mixed layer depth, or 0 if h1 < h0
}
\description{
Calculates rate of change in mixed layer depth for NCP estimation
returns positive value for MLD deepening i.e. Entrainment, otherwise returns 0
called by run_ncp functions
}
\keyword{depth}
\keyword{entrainment}
\keyword{layer}
\keyword{mixed}

