% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert_u10}
\alias{convert_u10}
\title{Correct wind speed to 10m height}
\usage{
convert_u10(U, z, method = "liutang")
}
\arguments{
\item{U}{numerical vector of wind speed}

\item{z}{height at x}

\item{method}{character string of either "liutang", "COARE" or "hsu". Default is "liutang"}
}
\value{
u10
}
\description{
Correct wind speed to 10m height
}
\details{
TODO
}
\references{
Liu W, Tang W. Equivalent neutral wind. Pasadena: Jet Propulsion Laboratory. 1996. Available: http://128.149.33.88/publication/paper/Liu-Tang-1996-jpl.pdf

COARE formulation courtesy of Tom Bell (Personal communication). 2013.

Hsu SA, Meindl EA, Gilhousen DB. Determining the Power-Law Wind-Profile Exponent under Near-Neutral Stability Conditions at Sea. J Appl Meteorol. 1994;33: 757-765
}

