% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solubility.R
\name{Csat.combined}
\alias{Csat.combined}
\title{Equilibrium Oxygen saturation concentration (combined fit)}
\usage{
Csat.combined(temp, salinity, unit = "mmolm")
}
\arguments{
\item{temp}{numeric vector of water temperature in degrees Celsius}

\item{salinity}{numeric vector of salinity (PSU)}

\item{unit}{"mmolm" for mmol m-3 (default), "mgl" for mg l-1 or "umolkg" for umol kg-1.}
}
\value{
vector of saturation concentration in mmol m-3
}
\description{
Calculates oxygen saturation concentration in equilibrium with the atmosphere
as per Garcia & Gordon, 1992, using the combined fit, which is not recommended.
This is however the formulation used on Aanderaa optodes.
}
\keyword{oxygen}
