% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Csat}
\alias{Csat}
\title{Oxygen concentration saturation}
\usage{
Csat(temp, salinity)
}
\arguments{
\item{temp}{numeric vector of water temperature in degrees Celsius}

\item{salinity}{numeric vector of salinity (PSU)}
}
\value{
vector of saturation concentration in micro mol per litre
}
\description{
Calculates oxygen saturation concentration in equilibrium with the atmosphere
as per Garcia & Gordon, 1992 (Benson & Kraus, rather than combined)
}
\examples{
Csat(10, 35)  # saturation at 10 degrees and 35 salinity
}
\keyword{oxygen}

