% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas.R
\name{dragCoef}
\alias{dragCoef}
\title{Drag Coefficient}
\usage{
dragCoef(u, method = "satLargePond")
}
\arguments{
\item{u}{vector of wind speed in meters per second at 10 meters height}

\item{method}{optional string determining parametrisation, valid strings are "satLargePond", "smith". default is "satLargePond"}
}
\value{
vector of drag coefficent
}
\description{
Calculates the drag coefficent
}
\details{
TODO, discuss methods
}
\references{
Sullivan, P. P., Romero, L., McWilliams, J. C., & Melville, W. K. (2012). Transient Evolution of Langmuir Turbulence in Ocean Boundary Layers Driven by Hurricane Winds and Waves. Journal of Physical Oceanography, 42(11), 1959-1980. doi:10.1175/JPO-D-12-025.1

Smith, S. D. (1980). Wind stress and heat flux over the ocean in gale force winds. J Phys. Oceanogr. 10, 709-726.
}
\keyword{coefficent}
\keyword{drag}

