% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gas.R
\name{kw}
\alias{kw}
\title{water side gas transfer velocity}
\usage{
kw(compound, T, u, S, method = "WA09", normalize = 0,
  schmidt_method = "mean")
}
\arguments{
\item{compound}{character string of compound of interest}

\item{T}{vector of temperature in degrees Centigrade}

\item{u}{vector of wind speed in meters per second at 10 meters height}

\item{S}{vector of salinity}

\item{method}{optional string determining kw parameterisation to use, see XXX for list default is 'WA09' (Wanninkhof et al, 2009)}

\item{normalize}{optional integer Schmitt number, if not 0 kw values are normalised to specified Schmitt number. Default is 0}

\item{schmidt_method,}{TODO. Default is 0}
}
\value{
vector of gas transfer velocity in meters per second
}
\description{
By default implements the gas transfer velocity parametrisation of Wanninkhov et al (2009).
}
\details{
TODO - K_calcs scheme
}
\examples{
kw('O2', 10, 7, 35)  # gas transfer velocity for oxygen at 10oC, 7 m-1 s-1 winds and 35 salinity.
}
\references{
Wanninkhof, R., Asher, W. E., Ho, D. T., Sweeney, C., & McGillis, W. R. (2009).  'Advances in Quantifying Air-Sea Gas Exchange and Environmental Forcing*. Annual Review of Marine Science, 1(1), 213-244. doi:10.1146/annurev.marine.010908.163742

TODO Nightingale et al 2000 (GBC)

TODO Wanninkhof et al 1992

TODO Wanninkhof 2014
}
\keyword{exchange}
\keyword{gas}
\keyword{kw}
\keyword{transfer}
\keyword{velocity}

