% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liang.R
\name{kw_L12}
\alias{kw_L12}
\title{Liang et al 2012 Kw param}
\usage{
kw_L12(u10, sst, sal = 35, gasname)
}
\arguments{
\item{u10}{}

\item{sst}{}

\item{sal}{}

\item{gasname}{}
}
\value{
vector of kw in meters per second
}
\description{
Liang et al 2012 Kw param
}
