% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gas.R
\name{LargePondCD}
\alias{LargePondCD}
\title{Saturated Large and Pond (1981) Drag Coefficent}
\usage{
LargePondCD(u)
}
\arguments{
\item{u}{vector of wind speed in meters per second at 10 meters height}
}
\value{
vector of drag coefficent
}
\description{
Calculates the drag coefficent of Large and Pond with the high wind saturation as per Sullivan et al, (2012)
}
\details{
TODO
}
\references{
Sullivan, P. P., Romero, L., McWilliams, J. C., & Melville, W. K. (2012). Transient Evolution of Langmuir Turbulence in Ocean Boundary Layers Driven by Hurricane Winds and Waves. Journal of Physical Oceanography, 42(11), 1959–1980. doi:10.1175/JPO-D-12-025.1
}
\keyword{coefficent}
\keyword{drag}

