% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/viscDen.R
\name{n_0}
\alias{n_0}
\title{LaLiberte viscosity}
\usage{
n_0(T)
}
\arguments{
\item{T}{vector of temperature in degrees Centigrade}
}
\value{
viscosity of pure water in cP (mPa s-1)
}
\description{
Viscosity of pure water according to LaLiberte 2007
}
\details{
TODO
}
\references{
LaLiberte 2007
}
\keyword{viscosity}

