% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscDen.R
\name{n_air}
\alias{n_air}
\title{Air viscosity}
\usage{
n_air(T)
}
\arguments{
\item{T}{vector of temperature in degrees Centigrade}
}
\value{
viscosity of saturated air TODO N.s/m^2 (Pa.s)
}
\description{
Dynamic Viscosity of Air
}
\details{
TODO
}
\references{
Tsiligiris, 2008
}
\keyword{air}
\keyword{viscosity}
