% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viscDen.R
\name{n_sw}
\alias{n_sw}
\title{Seawater viscosity}
\usage{
n_sw(T, S, return_relative_viscosity = FALSE)
}
\arguments{
\item{T}{vector of temperature in degrees Centigrade}

\item{S}{vector of salinity}

\item{return_relative_viscosity}{optional, if True returns relative viscosity (i.e. viscosity scaling factor), default is False.}
}
\value{
viscosity of seawater water in TODO kg/m/s
}
\description{
Viscosity of seawater
}
\details{
TODO
}
\references{
Hardy 1953

LaLiberte 2007
}
\keyword{seawater}
\keyword{viscosity}

