% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ncpModel}
\alias{ncpModel}
\title{NCP model}
\usage{
ncpModel(t, y, parms)
}
\arguments{
\item{t}{a integer vector timing sequence}

\item{y}{the initial oxygen concentration i.e. O2 at t0}

\item{parms}{a list of parameter values}
}
\value{
vector of predicted oxygen concentrations at solver timesteps
}
\description{
Solves for predicted oxygen for use in NCP estimation
Called by run_ncp functions.
}
\keyword{NCP}
\keyword{oxygen}

