% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncp.R
\docType{data}
\name{ncpTest}
\alias{ncpTest}
\title{NCP test data}
\format{a data frame with 53940 rows and 10 variables:
\describe{
\item{price}{usdollars}
\item{price}{usdollars}
}}
\source{
FIXME Tom
}
\usage{
ncpTest
}
\description{
example data for testing NCP calculations
}
\keyword{datasets}

