% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncp.R
\name{O2NCP.linear}
\alias{O2NCP.linear}
\title{O2 NCP}
\usage{
O2NCP.linear(dat, entrainment = T, kw_method = "WA13", asVolume = F)
}
\arguments{
\item{dat}{data frame matching the format outlined in XXX}

\item{entrainment}{if True (default) calculate NCP with entrainment, Cb0 and Cb1 must be supplied}

\item{kw_method}{character string passed to kw, default is 'WA13'}

\item{asVolume}{if True NCP is not multipled by average mixed layer depth}
}
\value{
a vector of NCP in mmol per m-3 per supplied time interval, or mmol per m-3 if asVolume is True.
}
\description{
calculate NCP based on O2 observations, assumes linear change in observed values between
timesteps
}
\details{
TODO
}

