% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncp.R
\name{O2NCP}
\alias{O2NCP}
\title{O2 NCP}
\usage{
O2NCP(x, entrainment = T)
}
\arguments{
\item{x}{data frame matching the format outlined in XXX}

\item{entrainment}{if True (default) calculate NCP with entrainment, Cb0 and Cb1 must be supplied}
}
\value{
a vector of NCP in uMol/l
}
\description{
calculate NCP based on O2 observations
}
\details{
TODO
}

