% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscDen.R
\name{p_air}
\alias{p_air}
\title{Air density}
\usage{
p_air(T)
}
\arguments{
\item{T}{vector of temperature in degrees Centigrade}
}
\value{
density of saturated air in kg m-3
}
\description{
density of saturated air according to Tsiligiris (2008)
}
\details{
TODO
}
\references{
Tsiligiris (2008)
}
\keyword{air}
\keyword{density}
