% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gas.R
\name{SmithCD}
\alias{SmithCD}
\title{Smith (1980) Drag Coefficient}
\usage{
SmithCD(u)
}
\arguments{
\item{u}{vector of wind speed in meters per second at 10 meters height}
}
\value{
vector of drag coefficent
}
\description{
Calculates the drag coefficent of Smith (1980)
}
\details{
TODO
}
\references{
Smith, S. D. (1980). Wind stress and heat flux over the ocean in gale force winds. J Phys. Oceanogr. 10, 709-726.
}
\keyword{coefficent}
\keyword{drag}

