% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscDen.R
\docType{data}
\name{sw_cmf}
\alias{sw_cmf}
\title{Mass fractions}
\format{
oA data frame with 53940 rows and 10 variables:
\describe{
\item{price}{usdollars}
\item{price}{usdollars}
}
}
\source{
FIXME Martin
}
\usage{
sw_cmf
}
\description{
Mass fractions
}
\keyword{datasets}
