% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncp.R
\name{transform_data}
\alias{transform_data}
\title{timeseries translator}
\usage{
transform_data(x)
}
\arguments{
\item{x}{a data frame matching the format outlined in XXX}
}
\value{
data.frame in the format expected by O2NCP()
}
\description{
divides up timeseries for NCP calculation
}
\details{
TODO
}

