% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscDen.R
\name{v_air}
\alias{v_air}
\title{Air dynamic viscosity}
\usage{
v_air(T)
}
\arguments{
\item{T}{vector of temperature in degrees Centigrade}
}
\value{
dynamic viscosity of seawater in cm/s
}
\description{
TODO calculate kinmatic viscosity of seawater in cm/s for Schmidt number calculation
}
\details{
TODO
}
\references{
TODO
}
\keyword{air}
\keyword{dynamic}
\keyword{viscosity}

