% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscDen.R
\name{v_sw}
\alias{v_sw}
\title{Seawater dynamic viscosity}
\usage{
v_sw(T, S)
}
\arguments{
\item{T}{vector of temperature in degrees Centigrade}

\item{S}{vector of salinity}
}
\value{
dynamic viscosity of seawater in cm/s
}
\description{
TODO calculate kinmatic viscosity of seawater in cm/s for Schmidt number calculation
}
\details{
TODO
}
\references{
TODO
}
\keyword{dynamic}
\keyword{seawater}
\keyword{viscosity}
