% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schmidt.R
\name{Vb}
\alias{Vb}
\title{molar volume}
\usage{
Vb(compound)
}
\arguments{
\item{compound}{character string}
}
\value{
vector of molar volume at boiling point
}
\description{
Calculates the molar volume at boiling point using the Schroeder method, or takes overruling value from compounds data.
note: for compounds containing elements other than C,H,O,N,S, Cl, Br, F, and I, the Schroeder method won't work so Vb must be specified
db, tb and rings should be the number of double and triple bonds, and ring features in the molecule respectively.
}
\details{
TODO
}
\references{
TODO
}
\keyword{volume}

