% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Wann09kw}
\alias{Wann09kw}
\title{Wanninkhov et al, 2009 Gas transfer velocity}
\usage{
Wann09kw(compound, T, u, S, normalize = 0, schmidt_formulation = 0)
}
\arguments{
\item{compound}{character string of compound of interest}

\item{T}{vector of temperature in degrees Centigrade}

\item{u}{vector of wind speed in meters per second at 10 meters height}

\item{S}{vector of salinity}
}
\value{
vector of gas transfer velocity in meters per second
}
\description{
Implements the gas transfer velocity parametrisation of Wanninkhov et al, 2009.
}
\examples{
Wann09kw('O2', 10, 7, 35)  # gas transfer velocity for oxygen
}
\references{
Wanninkhof, R., Asher, W. E., Ho, D. T., Sweeney, C., & McGillis, W. R. (2009).  'Advances in Quantifying Air-Sea Gas Exchange and Environmental Forcing*. Annual Review of Marine Science, 1(1), 213–244. doi:10.1146/annurev.marine.010908.163742
}
\keyword{exchange}
\keyword{gas}
\keyword{kw}

