% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wann14_coef}
\alias{wann14_coef}
\title{Coefficients for least squares fourth-order polynomial fits of Schmidt number vs temperature
For Salinity =35 as per Wanninkof 2014}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18 rows and 6 columns.
}
\usage{
wann14_coef
}
\description{
Coefficients for least squares fourth-order polynomial fits of Schmidt number vs temperature
For Salinity =35 as per Wanninkof 2014
}
\keyword{datasets}
