% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Bwoolf}
\alias{Bwoolf}
\title{Woolf and Thorpe, 1991 equilibrium bubble supersaturation}
\usage{
Bwoolf(compound = "O2", u)
}
\arguments{
\item{compound}{character string of compound of interest ('O2' ,'N2', 'Ar', 'CO2')}

\item{u}{vector of wind speed in meters per second at 10 meters height}
}
\value{
vector of equilibrium bubble supersaturation effects (no units)
}
\description{
Implements the bubble supersaturation parametrisation of Woolf and Thorpe, 1991
Note: parametrisation is only applicable for Oxygen (O2), Nitrogen (N2), Argon (Ar) and Carbon Dioxide (CO2).
}
\details{
TODO
}
\examples{
Bwoolf('Ar', 7)  # bubble supersaturation effect for Argon at 7 m/s winds
}
\references{
Woolf, D. K., & Thorpe, S. A. (1991). Bubbles and the air-sea exchange of gases in near-saturation conditions. Journal of Marine Research, 49(3), 435–466. doi:10.1357/002224091784995765
}
\keyword{bubble}
\keyword{exchange}
\keyword{gas}
\keyword{kw}

