% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schmidt.R
\docType{data}
\name{compounds}
\alias{compounds}
\title{compounds}
\format{oA data frame with 53940 rows and 10 variables:
\describe{
\item{price}{usdollars}
\item{price}{usdollars}
}}
\source{
FIXME Martin made it up
}
\usage{
compounds
}
\description{
compounds for Schmidt number calculations
}
\keyword{datasets}

