% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compounds}
\alias{compounds}
\title{compounds}
\format{
data frame with 53940 rows and 10 variables:
}
\usage{
compounds
}
\description{
compounds for Schmidt number calculations
}
\keyword{datasets}
