% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncp.R
\name{O2NCP.mean}
\alias{O2NCP.mean}
\title{Net community production (NCP) from O2}
\usage{
O2NCP.mean(
  dat,
  kw_method = "NG00",
  bubbles = T,
  entrainment = F,
  output_conc = F,
  debug = F
)
}
\arguments{
\item{dat}{data frame matching the format outlined in \link[airsea]{O2NCP.transform}}

\item{kw_method}{character string passed to kw, default is 'NG00'}

\item{bubbles}{logical value to enable bubble supersaturation param, default is TRUE (on)}

\item{entrainment}{logical value to enable entrainment calculation, default is FALSE (off)}

\item{output_conc}{logical value to enable expressing NCP as just a biological concentration change rather than a flux.}

\item{debug}{logical value to print debug messages
(units of mmol m-3), default is FALSE (off)}
}
\value{
a vector of NCP in mmol per m-3 per supplied time interval
}
\description{
calculate net community production based on O2 observations.
}
\details{
Full explanation of the method and methods are found in Hull et al,
  2015 Entrainment will only be calculated where an
  \code{entrainment} variable is found in the input data and it's value is
  \code{TRUE}. When being using within a monte-carlo analysis you can include
  bias estimates in the following variables \code{kw_error, B_error,
  Csat_error}. Where these variables are found these parameters will be
  scaled accordingly.
}
\references{
Hull et al, 2015 http://www.biogeosciences-discuss.net/12/15611/2015/bgd-12-15611-2015.html
}
