% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viscDen.R
\name{p_sw}
\alias{p_sw}
\title{Seawater density}
\usage{
p_sw(T, S)
}
\arguments{
\item{T}{vector of temperature in degrees Centigrade}

\item{S}{vector of salinity}
}
\value{
density of seawater in kg m-3
}
\description{
density of seawater according to Millero and Poisson (1981)
}
\details{
TODO
}
\references{
Millero and Poisson (1981)
}
\keyword{density}
\keyword{seawater}

