% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schmidt.R
\name{Sch}
\alias{Sch}
\title{Schmidt number}
\usage{
Sch(compound, T, S, method = "mean")
}
\arguments{
\item{compound}{character string}

\item{T}{vector of temperature in degrees Centigrade}

\item{S}{vector of salinity}

\item{method}{string matching Schmidt caculation method, default is 'mean'}
}
\value{
vector of molar volume at boiling point
}
\description{
Schmidt number
}
\details{
TODO
}
\references{
TODO
}
\keyword{volume}

