% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schmidt.R
\name{Sch}
\alias{Sch}
\title{Schmidt number}
\usage{
Sch(compound, TEMP, SAL, method = "JS")
}
\arguments{
\item{compound}{character string}

\item{TEMP}{vector of temperature in degrees Centigrade}

\item{SAL}{vector of salinity}

\item{method}{string matching Schmidt calculation method, default is 'WA'}
}
\value{
vector of molar volume at boiling point
}
\description{
Schmidt number
}
\details{
TODO
}
\references{
TODO
}
\keyword{volume}
