% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{transform_data}
\alias{transform_data}
\title{transform data to O2NCP format}
\usage{
transform_data(x)
}
\arguments{
\item{x}{data frame of observations}
}
\value{
data frame in correct format for O2NCP
}
\description{
transform data to O2NCP format
}
\details{
TODO
expects column headings to match the following format (order does not matter):
dateTime = POSIXct, T = temperature (oC), S = salinity, C = oxygen concentration (mmol m-3), u = wind speed (m s-1),
Pslp = pressure at sea level (mbar), h = mixed layer depth (m).
}

